Implemented the required machinery to support different serialization of data inside thunks.
This is done by wrapping the Thunk into a `Internal` new-type struct, which changes (de)serialization of the Thunk being done via the (De)Serialize implementation of the contained type, but rather by (de)serialization to a robtop-encoded string.

Note that this commit does not compile as these changes have no been implemented for LevelPassword. Since this branch is based off of an old version of the master branch, the implementation of (de)serialization for LevelPassword is still erroneous and we first need to rebase onto master to incorporate those changes before we can touch that code again.