Implemented (de)serialization for full levels
We serialize both partial and full levels into the same internal struct, which reduces code duplication greatly. The downside is that serde no longer catches incomplete responses, which are missing some fields in the full representation. I have not come across such a case yet (and it would be a bug on robtop's side), but everything's possible with this game's servers. For now, we treat the case where no level data was provided the same as a response without any of the extra data, and if one of the other values miss, we fill them in with dummy values.