Remove RefThunk
The complexity of this was not needed, as the "internal" version of each
struct does not accidentally need the ability to hold a borrowed
"Thunk::Processed" variant. It only needs to hold the unprocessed
string.

This commit changes the generated boilerplate code to do
"as_unprocessed" calls to convert Thunks into &str (previously, this
only happened in the Serialize implementation of RefThunk).