Use Dash-based API for Level structs
We cannot use #[derive(Dash)] here, because the derive macro does not
support the `LevelData` field. This is because we have two responses
(getGJLevel and downloadGJLevel) that map into the same struct, with the
former being a subset of the latter's indices. Supporting this would be
difficult, so for now just use a manual implementation of Dash here.