Add `Testable::canonicalize`
This plays the same role as `ThunkProcessor::process_all_thunks`, with
slightly more general semantics. If we do a save->load roundtrip, all
thunks needs to be evaluated at the end, as the "processed" artifacts
have all their thunks evaluated. If we do a load->save roundtrip, we
want to evaluate all thunks in the middle to ensure that evaluating
thunks does not break serializability.