Add feature flag to statically disable logging in serde module
Turns out that since this is a hot path, even with the log level set to
warn (e.g. no logs actually being emitted), this had a ~10% performance
impact.

Also get rid of the `last_token` field, I forgot that for non-hot-paths
(such as error reporting and logging), we already have the `nth_last`
function.