Add Thunk::as_processed
Prior to this, there was no non-mutating function implemented on `Thunk`
that allowed to retrieve a processed value (`process` modified the thunk
in-place, and `into_processed` consumed the thunk). The cost of this
ability is the requirement for `ThunkProcessor::Output`'s lifetime to
become covariant (as otherwise, the lifetimes of the Thunk::Processed
and Thunk::Unprocessed branches would be unreconcilable, with the former
being tied to &self, and the latter being 'a). This should not be a big
limitation in praxis though.