fix: allow decoding truncating base64
Apparently, the server data for Duelo Maestro got corrupted, and the
base64 string of the level description got truncated. Configure our
base64 decoder to not care about this, and simply decode as much as its
able to.

Of course, this means we cannot use Duelo Maestro for the roundtrip test
anymore, because our base64 _encoder_ will produce a valid base64 string
that definitely does not match the truncated input, so we'll need to
find a different level for this test.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>