fix: stop using dtoa for float serialization
While dtoa is a very efficient library for float->string conversions, it
does not match robtop's formatting exactly. Particularly, for floats
with no fractional part (e.g. integers such as '11'), it will append a
'.0' (e.g. format as '11.0'), while RobTop will format without decimal
part.

Thus, just use the standard library, which matches RobTop in this case.
It's slower, but since this is serialization, we do not care that much.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>