test: replace Duelo Maestro with The Badland
Dualo Maestro has malformed data on the server. While dash-rs will
happily process such malformed data as the GD servers would, it will not
_reproduce_ it during serialization (instead it will give an equivalent
representation that is well-formed and also understood by the GD
servers). As such, we cannot use this level for round trip testing
anymore. Replace it with 'The Badland'.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>