Change mod.rs to use OnceCell for a more convenient and dynamic way to edit URLs for GDPSes
The const was changed to a static of type ``OnceCell<&'static str>``.
``get_gd_url`` is now a call to ``OnceCell::get_or_init``, and the "or init" part is also set to the boomlings servers. (This is changeable, by the way. ;)). Also, i don't even know what i'm doing. So if i'm wrong please correct the mistakes i made in this commit.

Hope this helps,
yuptune